[Vectorize] Fix a warning
This patch fixes:

  llvm/lib/Transforms/Vectorize/LoadStoreVectorizer.cpp:1429:23:
  error: comparison of integers of different signs: 'int' and 'const
  size_t' (aka 'const unsigned long') [-Werror,-Wsign-compare]

+++ llvm/lib/Transforms/Vectorize/LoadStoreVectorizer.cpp
@@ -1426,7 +1426,7 @@ std::vector<Chain> Vectorizer::gatherChains(ArrayRef<Instruction *> Instrs) {
 
     bool MatchFound = false;
     auto ChainIter = MRU.begin();
-    for (int J = 0; J < MaxChainsToTry && ChainIter != MRU.end();
+    for (size_t J = 0; J < MaxChainsToTry && ChainIter != MRU.end();
          ++J, ++ChainIter) {
       std::optional<APInt> Offset =
           getConstantOffset(getLoadStorePointerOperand(ChainIter->first),
